this is a dev commit: 12 files changed, 225 insertions(+), 415 deletions(-)
 .github/xworkflowsx/tweet_burning_ship.yml  |   2 +-

 .github/xworkflowsx/tweet_mandelbrot.yml    |   2 +-

 __main__.py                                 |  11 +-

 core/fractal_burning_ship.py (new)          | 165 ++++++++++++++++++++++++++++

 core/get_ffmpeg_filter_vals.py (new)        |  12 ++

 core/save_img.py                            |  21 ++--

 core/shared.py                              |   2 +

 core/write_md.py (new)                      |  26 +++++

 tweet_burning_ship/__main__.py (gone)       | 135 -----------------------

 tweet_burning_ship/get_raw.py (gone)        |  52 ---------

 tweet_burning_ship/write_metadata.py (gone) | 106 ------------------

 tweet_mandelbrot/write_metadata.py (gone)   | 106 ------------------

message generated by QWD (github.com/nvfp/qwd)

+++ core/save_img.py
@@ -1,7 +1,7 @@
 import os, subprocess as sp
 from .shared import FFMPEG, RENDERED_IMG_PTH
 
-def save_img(
+def render_fractal_img(
     ppm_data,
 
     edit_contrast,
@@ -14,12 +14,6 @@ def save_img(
     edit_vignette,
     edit_temp
 ):
-    if os.path.exists(file_path):
-        raise FileExistsError(f'File already exists: {repr(file_path)}.')
-
-    if edit_vignette > 0: vig_mode = 'backward'
-    else: vig_mode = 'forward'
-
     filter = (
         'eq='
         f'contrast={edit_contrast}'
@@ -32,12 +26,11 @@ def save_img(
         
         ', vignette='
         f'a={abs(edit_vignette)}*PI/180'
-        f': mode={vig_mode}'
+        f": mode={'backward' if (edit_vignette > 0) else 'forward'}"
         
         ', colortemperature='
         f'temperature={edit_temp}'
     )
-
     cmd = [
         FFMPEG,
         '-v', 'error',
@@ -47,13 +40,17 @@ def save_img(
         '-i', '-',
         '-vf', filter,
         '-q:v', '1',
-        file_path
+        RENDERED_IMG_PTH
     ]
+    if os.path.exists(RENDERED_IMG_PTH): raise AssertionError
     pipe = sp.Popen(cmd, stdin=sp.PIPE)
     pipe.stdin.write(ppm_data)
     pipe.stdin.close()
     pipe.wait()
     pipe.terminate()
 
-    file_size = os.path.getsize(file_path)
-    return file_size  # metadata purposes
+    # file_size = os.path.getsize(file_path)
+    # return file_size  # metadata purposes
+
+def render_with_stats():
+    pass
